requirements.txt
streamlit>=1.44.0
requests>=2.32.0
pydantic>=2.10.0
python-dotenv>=1.0.1
openai>=1.75.0
pytest>=8.3.0
